requirements.txt
playwright
pyyaml
pydantic
python-dotenv
